folders and files refactor
