+++ constants/cod_qualificacao.py
