+++ constants/estados.py
