+++ constants/var.py
